Safari-611.1.10/LayoutTests/html5lib/generated/run-tests1-write-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token '<'
../resources/tests1.dat: PASS
